micro change

+++ ft_strncmp.c
@@ -6,15 +6,15 @@
 /*   By: rydelepi <rydelepi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/13 14:21:59 by rydelepi          #+#    #+#             */
-/*   Updated: 2025/10/20 13:36:54 by rydelepi         ###   ########.fr       */
+/*   Updated: 2025/10/23 15:13:00 by rydelepi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strncmp(const char *s1, const char *s2, unsigned int n)
+int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	if (n <= 0)
